Renamed getCoords.py to geocoding.py
